fix(build-studio): keep build-branch under module-size baseline for preserve guard
Move shouldPreserveBuildBranchWork into sandbox-source-currency and compact
the resume wiring so build-branch.ts shrinks (854 → 851) instead of growing.

Process-Spine-Decision: pure bug fix with regression tests; no new durable surface
